Fix enum deserialization (#234)
## Changes
In #230, enums were changed so that enum field names did not necessarily
match the enum value itself. However, the `_enum` helper method used
during deserialization of a response containing an enum was not updated
to handle this case. This PR corrects this method to check through the
values of the `__members__` of an enum, as opposed to the keys.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied