Support older versions of `urllib3` and Databricks Runtime with regards to `DEFAULT_METHOD_WHITELIST` change to `DEFAULT_ALLOWED_METHODS` (#240)
## Changes
This PR changes the SDK to tolerate urllib3 < 1.26.0. In these versions
of urllib3, Retry class's `DEFAULT_ALLOWED_METHODS` was called
`DEFAULT_METHOD_WHITELIST`, and the parameter in Retry's constructor was
changed from `method_whitelist` to `allowed_methods`. With this change,
users using older versions of DBR (like the 7.3 LTS) can still use the
current version of the SDK, even though the version of urllib3 is old.

This change can be reverted when we no longer need to support urllib3 <
1.26.0 (i.e. after DBR 10.4 LTS, which is EOL on [March 18,
2025](https://docs.databricks.com/release-notes/runtime/releases.html)).

## Test

- [x] Manually tested on DBR 7.3LTS:
![Old_DBR_test_-_Databricks](https://github.com/databricks/databricks-sdk-py/assets/1850319/62822f6c-5b6a-46f6-aa30-26584f2e7f6b)
- [x] Manually tested on DBR 13.1: 

![Old_DBR_test_-_Databricks_2](https://github.com/databricks/databricks-sdk-py/assets/1850319/da0f4ac9-5170-4fd7-bfee-662fcc9214a0)

Co-authored-by: Serge Smertin <259697+nfx@users.noreply.github.com>