fix(sql): workaround for filtering in query history api
This patch is intended to be a temporary workaround for the
inconsistency of the REST API [[1]]. The REST API states that all inputs
are query parameters, but in reality, they must be in the body.

This also works around the behavior of `filter_by` ignoring the
`page_token` when provided together by dropping the filter_by in
subsequent requests.

[1]: https://docs.databricks.com/api/workspace/queryhistory/list

Tested: Ran the Python SDK locally and confirmed I was able to filter
  and use the other parameters. Example code which filters by the
  query start time.

    from databricks.sdk import WorkspaceClient
    from databricks.sdk.service.sql import QueryFilter, QueryInfo, TimeRange
    w = WorkspaceClient(host=redacted, token=redacted)

    filter = QueryFilter(
        query_start_time_range=TimeRange(
            start_time_ms=1689626553081,
            end_time_ms=1689626553083,
        )
    )
    assert len(list(w.query_history.list(filter_by=filter))) == 1
    filter = QueryFilter(
        query_start_time_range=TimeRange(
            start_time_ms=1689626553083,
            end_time_ms=1689626553083,
        )
    )
    assert len(list(w.query_history.list(filter_by=filter))) == 0

Closes: https://github.com/databricks/databricks-sdk-py/issues/99
Signed-off-by: Luke Hsiao <luke.hsiao@numbersstation.ai>