Fixed serialization of lists of enum values (#248)
## Changes
Currently, lists of enum values in request objects are incorrectly
serialized in the `as_dict()` method. Instead of putting the enum
directly into the resulting dictionary, the enum's value needs to be
added to the dictionary.

Closes #214.

## Tests
Updated test_cluster_events to add a filter on EventType, which is a
list of enum values.