Remove `**kwargs` from methods and add argument validation
This PR removes `XxxRequest` classes that are referenced only as requests, effectively removing the undocumented `request=(XxxRequest(foo=1))` API. This will enforce validation on the field names for specific requests.

This PR also adds validation of custom types on the request level, so that it's clear also at the runtime level when an invalid request has been supplied.