Set necessary headers when authenticating via Azure CLI (#290)
## Changes
The Python SDK request authentication logic is inconsistent between the
Azure login types: for service principal auth, the SDK correctly adds
the X-Databricks-Azure-Workspace-Resource-Id when configured, but this
is missed for Azure CLI auth.

This PR fixes this by defining the logic to attach this header in a
common function that is used by all Azure-specific authentication types.

See https://github.com/databricks/databricks-sdk-go/pull/584 for the
same change in Go SDK.

## Tests
- [x] Added a unit test to ensure the header is being set for Azure CLI
login

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied