Apply attribute transformer when reading in attributes from the environment (#293)
## Changes
Attributes read from the environment are not currently parsed before
being stored in the `_inner` field of `Config`. As a result, the type of
the attribute will be different depending on whether the attribute is
read from the environment or passed to the constructor. This PR changes
all argument parsing to use `__setattr__` which converts the argument to
the expected type.

Closes #292.

## Tests
Unit test to ensure that string env vars are parsed to appropriate type
for the affected field.

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied