Propagate request headers to ApiClient (#302)
## Changes
Based on changes in
https://github.com/databricks/databricks-sdk-go/pull/572.

This PR adds initial support for headers in the ApiClient class. When
making a request, the SDK will construct a map of headers to be included
in the request. These are passed through a new parameter in the `do`
method.

In the future, to support dynamic headers, we can modify our codegen
template to add those fields to the headers map before calling the
ApiClient.

Alternatives considered:

* **Adding headers in request objects directly**: this would not require
any change in the ApiClient interface. Instead, we would introduce a new
annotation for headers, much like what we do for query parameters, and
reflectively scan the fields of each request object. We could do this,
but for headers that have a fixed value (like Content-Type and Accept),
these fields would essentially be private final fields that we could
iterate through. Users can never interact with these fields, so it seems
a bit unusual to have them in the request structure, a public interface.
## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied

---------

Signed-off-by: Miles Yucht <miles@databricks.com>