Fix test_files_api_read_twice_from_one_download (#326)
## Changes
This integration test is failing (and likely wasn't passing when I
merged the original PR). This PR fixes it by unifying the codepaths for
`__enter__` and `read` to call `_open`, which consistently will check if
the response is closed and throw an exception if so.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied