Don't try to import runtime_auth when not in runtime (#327)
## Changes
Runtime auth is attempted when we do not have any config variables set
(either directly in the sdk config, in .databrickcfg or as env vars).
This further attempts importing runtime_auth functions from the
dbruntime package. These are not available outside dbr.

This PR moves "is in dbr runtime" check before the import, so that the
import is not attempted outside dbr and users don't get cryptic errors.
## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied

---------

Signed-off-by: Kartik Gupta <88345179+kartikgupta-db@users.noreply.github.com>
Co-authored-by: Serge Smertin <259697+nfx@users.noreply.github.com>