Use `[]` instead of `None` as default list value
This PR makes experience consistent with Go SDK for the read path. At the moment, we use `None` as the default value for list fields. This may result in either unexpected runtime bugs, like

```
_backup = w.groups.get(...)
_backup_members = sorted([m.value for m in _backup.members])
...

E           TypeError: 'NoneType' object is not iterable
```

Technically, we can ask users to be very defensive in programming and do `sorted(m.value for m in _backup.members) if _backup.members else []`, but the most of the intended userbase does not know/want to know about these details.

What other pros and cons does keeping `None` as a default value?