Fix select spark version in staging (#388)
## Changes
In staging, there are some spark versions that do not have a patch
version. We fall back to a different pattern with no patch component for
these. Additionally, we check if the builds are equal to one another
(which allows us to avoid comparing one < the other in the case where
they are both None).

Closes #352 and #378.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] Added a unit test with a format that was previously not supported.
- [x] Hand-wrote an example using select_spark_versions(). This failed
on `main` but passed on my branch when targeting a staging workspace.