Make test_auth no longer fail if you have a default profile setup (#426)
## Changes
<!-- Summary of your changes that are easy to understand -->
- Previously `test_auth.py` will have several test cases failing if you
have a default profile locally, that's because the sdk code will always
look for the `~/.databrickscfg` file, which exists on your machine, but
is not expected
- Making specific test cases affected by this behavior use a fake file
system instead of the local disk to bypass this issue.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied