Update OpenAPI spec for Python SDK (#439)
## Changes
This PR updates the OpenAPI spec for the Python SDK to
d136ad0541f036372601bad9a4382db06c3c912d. As part of this, we add some
protection against fields or methods in our spec that conflict with
reserved keywords in Python.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied