Paginate all SCIM list requests in the SDK (#440)
## Changes
This PR incorporates two hard-coded changes for the SCIM API in the
Python SDK:
1. startIndex starts at 1 for SCIM APIs, not 0. However, the existing
.Pagination.Increment controls both the start index as well as whether
the pagination is per-page or per-resource. Later, we should replace
this extension with two independent OpenAPI options: `one_indexed`
(defaulting to `false`) and `pagination_basis` (defaulting to `resource`
but can be overridden to `page`).
2. If users don't specify a limit, the SDK will include a hard-coded
limit of 100 resources per request. We could add this to the OpenAPI
spec as an option `default_limit`, which is useful for any non-paginated
APIs that later expose pagination options and allow the SDK to
gracefully support those. However, we don't want to encourage folks to
use this pattern: all new list APIs are required to be paginated from
the start.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied

---------

Co-authored-by: Xinjie Zheng <xinjie.zheng@databricks.com>