Reference Documentation Refactoring (#467)
## Changes
This PR makes a number of changes to the documentation with the aim of
improving the coverage of reference material as well as making the
reference documentation more intuitive to follow by mirroring the
structure of the library.

The biggest change here is the architecture: we're no longer generating
ReStrucutredText by hand, but rather using the `automodule` directive to
generate most of the documentation for us. This means that all classes
and enums automatically have documentation, type annotations are present
for all parameters, and links and references work properly. The main
downside is that examples are no longer included in the main
documentation, so I've added a page to the top-level TOC to point out
the examples in our repo directly.

To improve the documentation coverage of existing classes, I tweaked
some of our templates to include docs from the OpenAPI spec directly in
the WorkspaceClient/AccountClient definitions, as well as in the
generated code for dataclasses and enums. There are many things that are
poorly documented now; we'll need to lean on better linting in the
upstream OpenAPI spec to prevent this from getting worse.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied