Install local library when generating docs (#469)
## Changes
#467 changed the docs to require the databricks SDK itself to be
installed for Sphinx to generate docs correctly. This PR fixes the
requirements file used by readthedocs.org to install the SDK before
trying to generate docs. The requirements file is consumed from the root
directory by readthedocs.py, so I moved it to the root directory so it's
more obvious that it ought to be run from there.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied