Fix readme links in pypi (#472)
## Changes
This PR fixes readme links so that they work in PyPI, according to this
comment: https://github.com/pypa/readme_renderer/issues/169.

I've also replaced the examples section of the readme to match what we
have in our documentation, which may be easier for folks to follow.

Resolves https://github.com/databricks/databricks-sdk-py/issues/289.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied