Sort imports in service template (#479)
## Changes
Sort imports in service template. The following line must be the first
command in a Python file:
`from __future__ import annotations` 
Usually this is sorted by the linter. But if the file generated is not
syntactically correct, it won't be sorted and it will cause a error
which will mask the real error with the generated file.

## Tests
`deco openapi generate-sdk py --openapi-spec master`

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied