Support dev and staging workspaces (#514)
## Changes
Copies the DatabricksEnvironment abstraction from the Go SDK to the
Python SDK. This enables using the SDK for Azure development and staging
workspaces, which use an alternate Azure login application ID.

Additionally, this PR changes `azure_environment` to accept the same
values as the Go SDK and Terraform providers for both Databricks and
Azure, i.e. `PUBLIC`, `USGOVERNMENT`, and `CHINA`.

The first commit of this PR is a refactor moving config & credentials
provider code to separate files, as core.py is getting quite big and
difficult to maintain. The second one is much smaller and has just the
net new bits.

## Tests
- [x] Unit test for the new environment lookup code.
- [x] Manual test Azure CLI auth against a staging Azure workspace.