Make docs useful again (#519)
## Changes
The last refactor to docs introduced dataclasses and enum documentation
but regressed on examples and ease of use/navigability by introducing
extremely long pages with many different APIs and classes documented.
This has led to some confusion for users of the SDK who are simply
trying to explore/understand what functionality is supported.

This PR undoes that refactor while preserving the dataclass and enum
documentation, and it even improves on it somewhat, reducing a
substantial amount of visual noise from dataclass constructor signatures
and making enum values easier to read and understand.

Link to preview:
https://databricks-sdk-py--519.org.readthedocs.build/en/519/

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied