Use fake clock for faster unit tests (#533)
## Changes
Some unit tests of retry logic are slow because they sleep for a second
or two. This PR changes ApiClient to accept a Clock which supplies the
time for retries and waiting. In unit tests, we use a fake clock that
doesn't tick until sleep() is called, and then ticks by the amount
provided.

## Tests
Existing unit tests.

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied