Add get_workspace_id to WorkspaceClient (#537)
## Changes
There are times when it is especially useful to get the workspace ID for
the current workspace client. Currently, the workspace ID for the
current workspace is exposed as a header in the SCIM Me API call. We'll
expose this through a get_workspace_id() method, caching the workspace
ID for the lifetime of the client.

In the future, we may add a meta service for exposing information about
the current account/workspace. At that point, we can migrate off of this
somewhat hacky approach.

Ports https://github.com/databricks/databricks-sdk-go/pull/808 to the
Python SDK.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied