Make a best effort attempt to initialise all Databricks globals (#562)
## Changes
We only initialise dbutils locally when using `from
databricks.sdk.runtime import *`. Users in the webui are guided to use
this import in all library code.
<img width="812" alt="image"
src="https://github.com/databricks/databricks-sdk-py/assets/88345179/3f28bc3c-0ba9-41ac-b990-3c4f5bf138aa">

The local (for people outside DBR) solution so far was to initialise
spark manually. But this can be tedious for deeply nested libraries
(which is the reason this import was introduced in the first place).

Now, we make a best effort attempt to initialise maximum number of
globals locally, so that users can build and debug libraries using
databricks connect.

## Tests
* integration test

- [ ] `make test` run locally
- [x] `make fmt` applied
- [ ] relevant integration tests applied

---------

Signed-off-by: Kartik Gupta <88345179+kartikgupta-db@users.noreply.github.com>
Co-authored-by: Miles Yucht <miles@databricks.com>