New example to list compute resource for SUBMIT_RUN job runs (#572)
## Changes
From a discussion with @mgyucht on a request for getting a list of jobs
submitted via jobs/submit and the associated compute resource(s). There
are multiple compute-related fields returned in the response (both on
the run, and on the task) so this example aims to disambiguate them.

Ran against AWS, Azure, GCP workspaces.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied