Properly escape multi-segment path parameters (#596)
## Changes
Ports https://github.com/databricks/databricks-sdk-go/pull/869 to Python
SDK. Subsumes #592.

Currently, path parameters are directly interpolated into the request
URL without escaping. This means that characters like /, ? and # will
not be percent-encoded and will affect the semantics of the URL,
starting a new path segment, query parameters, or fragment,
respectively. This means that it is impossible for users of the Files
API to upload/download objects that contain ? or # in their name. / is
allowed in the path of the Files API, so it does not need to be escaped.

The Files API is currently marked with x-databricks-multi-segment,
indicating that it should be permitted to have / characters but other
characters need to be percent-encoded. This PR implements this.

## Tests

- [x] Unit test for multi-segment path escaping behavior.
- [x] Updated integration test to use # and ? symbols in the file name.
This failed on main but succeeded on this PR.