DBUtils implementation for Volumes (#623)
## Changes
UC Volumes has been released for some time, but users are unable to use
UC Volumes with `dbutils.fs` from the SDK. This PR implements support
for `/Volumes` paths in DBUtils in the SDK.

I've done this primarily by extending `DbfsExt` to work with UC Volumes
paths. A new class `_VolumePath` is supported, implementing the set of
base operations defined on the `_Path` parent abstract class for volume
paths. An accompanying `_VolumesIO` is also implemented to provide a
consistent interface for reading from and writing to UC Volumes files,
especially for writing: the `download` API already returns a `BinaryIO`,
but the `upload` API _accepts_ a `BinaryIO`, so this adapter allows for
a user to "open" a Volumes path for writing.

In order to properly implement `ls`, I changed the existing
implementation of `list` in `_Path` subclasses to return a generator of
`FileInfo`s. This allows for better reuse of this common functionality
between `ls`, `cp` and `mv`.

Open questions:
* Should we do anything to distinguish between a path named `/Volumes`
on disk and UC Volumes? E.g. via a different scheme?

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied