Remove unnecessary `dbfs:` scheme when listing in `w.dbutils.fs.ls()` (#631)
## Changes
#623 introduced DBUtils support for volumes but also caused a small
regression in listing behavior: `dbutils.fs.ls()` should not include the
`dbfs:` scheme. This PR makes that fix. Additionally, it fixes a small
bug in volumes recursive listing, only including the file paths as is
the behavior with DBFS.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied