Better error message for users trying to use `abfss` or `s3` schemes with SDK DBUtils (#634)
## Changes
DBUtils in the SDK only supports the `dbfs` and `file` schemes. Within
DBFS, only paths in the root filesystem are supported, not mountpoints.
Other schemes are not supported. To minimize customer confusion, we can
check for this case during all DBUtils FS operations when constructing
the path.

## Tests
Added a unit test for this case.

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied