Add `serverless_compute_id` field to the config (#685)
## Changes
We propose adding a new `serverless_compute_id` attribute in the config
that will be used to enable serverless compute in the downstream
applications. The default value to enable serverless is
`serverless_compute_id = auto`, other values might be used in the future
if needed. Config doesn't have to validate the value in
`serverless_compute_id`, this will be done in the downstream
applications.

## Tests

- [x] `make test` run locally
- [x] `make fmt` applied
- [ ] relevant integration tests applied