Support partners in SDK (#648)
## Changes
Ports https://github.com/databricks/databricks-sdk-go/pull/925 to the
Python SDK.

Partners of Databricks need a mechanism to register themselves in
libraries or applications that they write. In this way, requests made by
users of those libraries will include sufficient information to link
those requests to the original users.

This PR adds a new `useragent` module with functions to manipulate the
user agent.
* `product()`: returns the globally configured product & version.
* `with_product(product: str, product_version: str)`: configure the
global product & version.
* `extra()`: returns the globally configured extra user agent metadata.
* `with_extra(key: str, value: str)`: add an extra entry to the global
extra user agent metadata.
* `with_partner(partner: str)`: add a partner to the global extra user
agent metadata.
* `to_string(product_override: Optional[Tuple[str, str]]=None,
other_info: Optional[List[Tuple[str, str]]] = None): str`: return the
User-Agent header as a string.


One new function here is `with_partner`, which can be used by a partner
to add partner information to the User-Agent header for requests made by
the SDK. The new header will have the form `partner/<parther id>`. The
partner identifier is opaque for the SDK, but it must be alphanumeric.

This PR also removes the requirement that a user agent entry contain
only a single copy of each key. This allows multiple partners to
register in the same library or application.

In this PR, I've also refactored the user agent library to be more
static, aligning it with the Go and Java SDKs. This makes it easier to
maintain and ensures similar behavior between all 3 SDKs. Note that this
SDK has extra functionality that doesn't exist in the Go and Java SDKs,
namely config-level user agent info; that is preserved here.

## Tests
Unit tests were added to verify that the user agent contains all
expected parts and supports multiple partners.

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied