[Fix] Decrease runtime of recursive workspace listing test (#721)
## Changes
The current integration test for recursive workspace listing is very
slow because it lists all resources in a very large directory (the
integration test user's home folder). To decrease the time this test
takes, we can simply create a directory with a file and a subdirectory
with another file. This means the test requires only two API calls to
complete.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied