[Internal] Fix get_workspace_client test to match Go SDK behavior (#738)
## Changes
The current get_workspace_client test fails because the SP used by the
test does not have access to the first workspace listed. In the
[Go](https://github.com/databricks/databricks-sdk-go/blob/main/internal/account_client_test.go#L12)
&
[Java](https://github.com/databricks/databricks-sdk-java/blob/1b90e2318f8221ac0a6e4b56c9b0e4c286e38c9f/databricks-sdk-java/src/test/java/com/databricks/sdk/integration/AccountClientIT.java#L17)
SDKs, the corresponding test respects the `TEST_WORKSPACE_ID`
environment variable to know which workspace to attempt to login to.
This PR changes the test to use that environment variable as well.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied