[Fix] Handle non-JSON errors gracefully (#741)
## Changes
Some errors returned by the platform are not serialized using JSON (see
https://github.com/databricks/databricks-sdk-go/issues/998 for an
example). They are instead serialized in the form "<ERROR_CODE>:
<MESSAGE>". Today, the SDK cannot parse these error messages well,
resulting in a poor user experience.

This PR adds support for parsing these error messages from the platform
to the SDK. This should reduce bug reports for the SDK with respect to
unexpected response parsing. This PR also refactors the error
deserialization logic somewhat to make it more extensible in the future
for other potential error formats that are not currently handled.

As a side-effect of this change, I've refactored the structure of the
error handling in the Python SDK to more closely reflect how errors are
handled in the Go SDK. This should make maintenance more straightforward
in the future. It also introduces a new error message to the Python SDK
to refer users to our issue tracker when the SDK receives an error
response that it cannot parse, like what we do in the Go SDK.

Ports https://github.com/databricks/databricks-sdk-go/pull/1031 to the
Python SDK.

## Deprecations
This PR deprecates several fields in the constructor for
DatabricksError. Going forward, SCIM-specific and API 1.2-specific
parameters should not be specified in the constructor; instead, they
will be handled in error parsers.

## Breaking Changes
The introduction of a different message for non-JSON responses may be a
breaking change if users matched on the message structure used before.

## Tests
Existing tests still pass, adding tests before merging this.

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied